feat: Add secondaryIconColor to theme
This commit adds the secondaryIconColor to the theme constant file in order to define a secondary color for icons. This color will be used for certain icons in the application.

Refactor and cleanup code

This commit removes unused code related to production code properties and items. The unused code is no longer necessary and has been safely removed.

feat: Add ProductionCodePropertyItemResponse interface

This commit adds the ProductionCodePropertyItemResponse interface, which defines the structure of a response object for production code property items. This interface will be used in the application to handle and manipulate data related to production code properties.

feat: Update BaseUrl to use new server IP address

The BaseUrl constant in the API index file has been updated to use the new server IP address. This change ensures that the application communicates with the correct server. The new BaseUrl is "http://localhost:8080".

feat: Add green color to theme

This commit adds the green color to the theme constant file. The green color will be used to style certain elements in the application's theme.